% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations_test}
\alias{stations_test}
\title{Station Locations Test Data}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{station_id}{Station identifier (character)}
\item{name}{Station name (character)}
\item{longitude}{Longitude in decimal degrees (numeric)}
\item{latitude}{Latitude in decimal degrees (numeric)}
}
}
\source{
Simulated data for package testing
}
\usage{
stations_test
}
\description{
A sample dataset containing GPS coordinates of 4 weather stations
in the Souss-Massa region of Morocco.
}
\keyword{datasets}
