% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather_test}
\alias{weather_test}
\title{Weather Test Data}
\format{
A data frame with 48 rows and 6 variables:
\describe{
\item{date}{Date of measurement (Date)}
\item{temperature}{Air temperature in Celsius (numeric)}
\item{humidity}{Relative humidity in percent (numeric)}
\item{wind_speed}{Wind speed in m/s (numeric)}
\item{radiation}{Solar radiation in W/m2 (numeric)}
\item{station_id}{Station identifier (character)}
}
}
\source{
Simulated data for package testing
}
\usage{
weather_test
}
\description{
A sample weather dataset containing monthly meteorological measurements
for 4 stations in the Souss-Massa region of Morocco (2023).
}
\keyword{datasets}
